+++ backend/canisters/local_user_index/api/src/lifecycle/init.rs
@@ -14,5 +14,6 @@ pub struct Args {
     pub group_index_canister_id: CanisterId,
     pub notifications_canister_id: CanisterId,
     pub cycles_dispenser_canister_id: CanisterId,
+    pub internet_identity_canister_id: CanisterId,
     pub test_mode: bool,
 }