+++ backend/canisters/local_user_index/impl/src/lifecycle/inspect_message.rs
@@ -17,6 +17,7 @@ fn accept_if_valid(runtime_state: &RuntimeState) {
 
     let is_valid = match method_name.as_str() {
         "join_group" | "report_message" | "invite_users_to_group" => runtime_state.is_caller_openchat_user(),
+        "register_user" => true,
         _ => false,
     };