+++ backend/canisters/user_index/impl/src/model/pending_payments_queue.rs
@@ -35,5 +35,4 @@ pub struct PendingPayment {
 pub enum PendingPaymentReason {
     Treasury,
     ReferralReward,
-    BitcoinMiamiReferral,
 }