+++ backend/canisters/user_index/impl/src/queries/check_username.rs
@@ -1,7 +1,7 @@
-use crate::updates::set_username::{validate_username, UsernameValidationResult};
 use crate::{read_state, RuntimeState};
 use ic_cdk_macros::query;
 use user_index_canister::check_username::{Response::*, *};
+use utils::username_validation::{validate_username, UsernameValidationError};
 
 #[query]
 fn check_username(args: Args) -> Response {
@@ -10,15 +10,15 @@ fn check_username(args: Args) -> Response {
 
 fn check_username_impl(args: Args, runtime_state: &RuntimeState) -> Response {
     match validate_username(&args.username) {
-        UsernameValidationResult::Ok => {
+        Ok(_) => {
             if runtime_state.data.users.does_username_exist(&args.username) {
                 UsernameTaken
             } else {
                 Success
             }
         }
-        UsernameValidationResult::TooShort(min_length) => UsernameTooShort(min_length),
-        UsernameValidationResult::TooLong(max_length) => UsernameTooLong(max_length),
-        UsernameValidationResult::Invalid => UsernameInvalid,
+        Err(UsernameValidationError::TooShort(min_length)) => UsernameTooShort(min_length),
+        Err(UsernameValidationError::TooLong(max_length)) => UsernameTooLong(max_length),
+        Err(UsernameValidationError::Invalid) => UsernameInvalid,
     }
 }