+++ frontend/openchat-agent/src/services/localUserIndex/candid/idl.d.ts
@@ -1,6 +1,7 @@
 import type { IDL } from "@dfinity/candid";
 import {
     _SERVICE,
+    RegisterUserResponse,
     GroupCanisterGroupChatSummary,
     JoinGroupResponse,
     InviteUsersToGroupResponse,
@@ -9,6 +10,7 @@ import {
 } from "./types";
 export {
     _SERVICE as LocalUserIndexService,
+    RegisterUserResponse as ApiRegisterUserResponse,
     GroupCanisterGroupChatSummary as ApiGroupCanisterGroupChatSummary,
     JoinGroupResponse as ApiJoinGroupResponse,
     GateCheckFailedReason as ApiGateCheckFailedReason,