+++ frontend/openchat-agent/src/services/openchatAgent.ts
@@ -1316,6 +1316,8 @@ export class OpenChatAgent extends EventTarget {
         username: string,
         referralCode: string | undefined
     ): Promise<RegisterUserResponse> {
+        // const localUserIndex = await this._userIndexClient.userRegistrationCanister();
+        // return this.createLocalUserIndexClient(localUserIndex).registerUser(username, referralCode);
         return this._userIndexClient.registerUser(username, referralCode);
     }